3.0.4

+++ bot.js
@@ -124,7 +124,7 @@ client.on("guildCreate", async (guild) => {
             VALUES('${guild.id}')
             `
         );
-            
+
         guildLocalPrefixes.set(guild.id, config.PREFIX.LOCAL);
 
         console.log(colors.brightYellow(`${getDateTime()} >>> Joined guild: ${guild.name} (ID: ${guild.id})`));
@@ -160,7 +160,7 @@ client.on("guildDelete", async (guild) => {
             WHERE guildId = '${guild.id}'
             `
         );
-            
+
         guildLocalPrefixes.delete(guild.id);
 
         console.log(colors.brightYellow(`${getDateTime()} >>> Left guild: ${guild.name} (ID: ${guild.id})`));
@@ -865,9 +865,9 @@ async function COMMAND_UNSETUP(sentMessage) {
         WHERE guildId = '${sentMessage.message.guild.id}'
         `
     );
-    
-    sentMessage.message.guild.channels.cache.get(channelCreator).delete();
-    sentMessage.message.guild.channels.cache.get(channelCreatorCategory).delete();
+
+    if (sentMessage.message.guild.channels.cache.find((c) => c.id === channelCreator)) sentMessage.message.guild.channels.cache.get(channelCreator).delete();
+    if (sentMessage.message.guild.channels.cache.find((c) => c.id === channelCreatorCategory)) sentMessage.message.guild.channels.cache.get(channelCreatorCategory).delete();
 
     consoleLoggingCommands(sentMessage, "SUCCESS");
